Revert "chore: change Decimal to int for calc_allocations & align tests"
This reverts commit 7561c8999c59ad9c806a54105e19aed7ce3001ad.

+++ backend/app/engine/projects/rewards/allocations/quadratic_funding.py
@@ -7,10 +7,10 @@ from app.engine.projects.rewards import ProjectAllocations
 
 
 class QuadraticFundingAllocations(ProjectAllocations):
-    def _calc_allocations(self, allocations: List[AllocationItem]) -> int:
+    def _calc_allocations(self, allocations: List[AllocationItem]) -> Decimal:
         sum_sqrts = sum(
             Decimal(sqrt(allocation.uq_score * allocation.amount))
             for allocation in allocations
         )
         quadratic_value = sum_sqrts**2
-        return int(quadratic_value)
+        return quadratic_value
